+++ src/mock/movie.js
@@ -1,17 +1,32 @@
 import dayjs from 'dayjs';
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
+import {getRandomInteger, getRandomFloat} from '../utils.js';
 
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomFloat = (a = 1, b = 0) => {
-  const lower = Math.min(a, b);
-  const upper = Math.max(a, b);
-  return lower + Math.random() * (upper - lower);
-};
+const titles = [
+  'Made for Each Other',
+  'The Dance of Life',
+  'Sagebrush Trail',
+  'The Man with the Golden Arm',
+  'Santa Claus Conquers the Martians',
+  'Popeye the Sailor Meets Sindbad the Sailor',
+  'The Great Flamarion',
+];
+const images = [
+  'made-for-each-other.png',
+  'popeye-meets-sinbad.png',
+  'sagebrush-trail.jpg',
+  'santa-claus-conquers-the-martians.jpg',
+  'the-dance-of-life.jpg',
+  'the-great-flamarion.jpg',
+  'the-man-with-the-golden-arm.jpg',
+];
+const genres = [
+  'Musical',
+  'Western',
+  'Drama',
+  'Comedy',
+  'Cartoon',
+  'Mystery',
+];
 
 const generateDescription = () => {
   const descriptions = [
@@ -38,20 +53,10 @@ const generateDescription = () => {
   return description.trim();
 };
 
-const generateTitle = () => {
-  const titles = [
-    'Made for Each Other',
-    'The Dance of Life',
-    'Sagebrush Trail',
-    'The Man with the Golden Arm',
-    'Santa Claus Conquers the Martians',
-    'Popeye the Sailor Meets Sindbad the Sailor',
-    'The Great Flamarion',
-  ];
+const generateRandomData = (values) => {
+  const randomIndex = getRandomInteger(0, values.length - 1);
 
-  const randomIndex = getRandomInteger(0, titles.length - 1);
-
-  return titles[randomIndex];
+  return values[randomIndex];
 };
 
 const generateRating = () => {
@@ -60,42 +65,11 @@ const generateRating = () => {
   return randomRating.toFixed(2);
 };
 
-const generatePoster = () => {
-  const images = [
-    'made-for-each-other.png',
-    'popeye-meets-sinbad.png',
-    'sagebrush-trail.jpg',
-    'santa-claus-conquers-the-martians.jpg',
-    'the-dance-of-life.jpg',
-    'the-great-flamarion.jpg',
-    'the-man-with-the-golden-arm.jpg',
-  ];
-
-  const randomIndex = getRandomInteger(0, images.length - 1);
-
-  return images[randomIndex];
-};
-
 const generateDuration = () => {
   const hour = getRandomInteger(0, 1);
   const min = getRandomInteger(10, 60);
 
   return hour ? `${hour}h ${min}m` : `${min}m`;
-};
-
-const generateGenre = () => {
-  const genres = [
-    'Musical',
-    'Western',
-    'Drama',
-    'Comedy',
-    'Cartoon',
-    'Mystery',
-  ];
-
-  const randomIndex = getRandomInteger(0, genres.length - 1);
-
-  return genres[randomIndex];
 };
 
 const generateDate = () => {
@@ -111,22 +85,22 @@ export const generateMovie = () => {
   const genresCount = getRandomInteger(0, 3);
 
   return {
-    poster: generatePoster(),
-    title: generateTitle(),
+    poster: generateRandomData(images),
+    title: generateRandomData(titles),
     rating: generateRating(),
     releaseYear: getRandomInteger(1900, 1970),
     duration: generateDuration(),
-    genre: generateGenre(),
+    genre: generateRandomData(genres),
     description: generateDescription(),
     commentsCount: getRandomInteger(0, 5),
-    fullPoster: generatePoster(),
-    originalTitle: generateTitle(),
+    fullPoster: generateRandomData(images),
+    originalTitle: generateRandomData(titles),
     director: 'Robert B. Weide',
     writers: 'Anne Wigton, Heinz Herald, Richard Weil',
     actors: 'Erich von Stroheim, Mary Beth Hughes, Dan Duryea',
     releaseDate: generateDate(),
     country: 'USA',
-    genres: new Array(genresCount).fill().map(generateGenre),
+    genres: new Array(genresCount).fill().map(() => generateRandomData(genres)),
     fullDescription: 'The film opens following a murder at a cabaret in Mexico City in 1936, and then presents the events leading up to it in flashback. The Great Flamarion (Erich von Stroheim) is an arrogant, friendless, and misogynous marksman who displays his trick gunshot act in the vaudeville circuit. His show features a beautiful assistant, Connie (Mary Beth Hughes) and her drunken husband Al (Dan Duryea), Flamarion`s other assistant. Flamarion falls in love with Connie, the movie`s femme fatale, and is soon manipulated by her into killing her no good husband during one of their acts.',
     ageRating: getRandomInteger(0, 21),
     isToWatch: Boolean(getRandomInteger(0, 1)),
